update(printer): move printer package from internal to use KICS as a module
