+++ pkg/printer/utils.go
